+++ MEWMIProv/Src/UPID_Commands.cpp
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 /*
- * Copyright (C) 2020-2022 Intel Corporation
+ * Copyright (C) 2020-2023 Intel Corporation
  */
 /*++
 
@@ -94,7 +94,7 @@ uint32_t UPID_Commands::GetUPIDCommand(uint32_t& oemPlatformIdType, std::wstring
 		}
 		oemPlatformId.assign(oemPlatformIdStr.str());
 
-		for (int i = 0; i < UPID_LEN; i++) //csmePlatformId is always Binary - Show it as hex string
+		for (size_t i = 0; i < UPID_LEN; i++) //csmePlatformId is always Binary - Show it as hex string
 		{
 			csmePlatformIdStr << "0x" << std::hex << std::setfill(L'0') << std::setw(2) << unsigned(response.CSMEPlatformId[i]) << " ";
 		}
